Moves the non-VS specific bits to the Managed project for ActiveEditorContextTracker.
