+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS/ProjectSystem/VS/LanguageServices/ActiveEditorContextTracker.cs
@@ -1,36 +1,32 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
 
 using Microsoft.VisualStudio.ProjectSystem.LanguageServices;
-using Microsoft.VisualStudio.ProjectSystem.Utilities;
 using Microsoft.VisualStudio.TextManager.Interop;
-using Microsoft.VisualStudio.Threading;
 using HierarchyId = Microsoft.VisualStudio.Shell.HierarchyId;
 
 namespace Microsoft.VisualStudio.ProjectSystem.VS.LanguageServices
 {
     /// <summary>
-    ///     Provides an implementation of <see cref="IActiveEditorContextTracker"/> that tracks the
-    ///     "active" context for the editor by handling the VSHPROPID_ActiveIntellisenseProjectContext
-    ///     hierarchy property.
+    ///     Tracks the "active" context for the editor by handling the VSHPROPID_ActiveIntellisenseProjectContext hierarchy property.
     /// </summary>
     [Export(typeof(IActiveIntellisenseProjectProvider))]
-    [Export(typeof(IActiveEditorContextTracker))]
     [ExportProjectNodeComService(typeof(IVsContainedLanguageProjectNameProvider))]
     [AppliesTo(ProjectCapability.DotNetLanguageService)]
-    internal class ActiveEditorContextTracker : IActiveIntellisenseProjectProvider, IVsContainedLanguageProjectNameProvider, IActiveEditorContextTracker
+    internal class VsActiveEditorContextTracker : IActiveIntellisenseProjectProvider, IVsContainedLanguageProjectNameProvider
     {
-        private ImmutableList<string> _contexts = ImmutableList<string>.Empty;
-        private string? _activeIntellisenseProjectContext;
+        private readonly IActiveEditorContextTracker _activeEditorContextTracker;
 
+        // UnconfiguredProject is only included for scoping reasons.
         [ImportingConstructor]
-        public ActiveEditorContextTracker(UnconfiguredProject? project) // For scoping
+        public VsActiveEditorContextTracker(UnconfiguredProject? project, IActiveEditorContextTracker activeEditorContextTracker)
         {
+            _activeEditorContextTracker = activeEditorContextTracker;
         }
 
         public string? ActiveIntellisenseProjectContext
         {
-            get { return _activeIntellisenseProjectContext ?? _contexts.FirstOrDefault(); }
-            set { _activeIntellisenseProjectContext = value; }
+            get { return _activeEditorContextTracker.ActiveIntellisenseProjectContext; }
+            set { _activeEditorContextTracker.ActiveIntellisenseProjectContext = value; }
         }
 
         public int GetProjectName(uint itemid, out string? pbstrProjectName)
@@ -46,35 +42,10 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.LanguageServices
             return HResult.OK;
         }
 
-        public bool IsActiveEditorContext(string contextId)
-        {
-            Requires.NotNullOrEmpty(contextId, nameof(contextId));
+        // Pass-through for unit testing.
+        internal bool IsActiveEditorContext(string contextId) => _activeEditorContextTracker.IsActiveEditorContext(contextId);
 
-            if (!_contexts.Contains(contextId))
-                throw new InvalidOperationException($"Context with ID '{contextId}' has not been registered or has already been unregistered.");
-
-            return StringComparers.WorkspaceProjectContextIds.Equals(ActiveIntellisenseProjectContext, contextId);
-        }
-
-        public IDisposable RegisterContext(string contextId)
-        {
-            Requires.NotNullOrEmpty(contextId, nameof(contextId));
-
-            bool changed = ThreadingTools.ApplyChangeOptimistically(ref _contexts, projectContexts =>
-            {
-                if (!projectContexts.Contains(contextId))
-                {
-                    projectContexts = projectContexts.Add(contextId);
-                }
-
-                return projectContexts;
-            });
-
-            if (!changed)
-                throw new InvalidOperationException($"Context with ID '{contextId}' has already been registered.");
-
-            return new DisposableDelegate(
-                () => Assumes.True(ThreadingTools.ApplyChangeOptimistically(ref _contexts, contextId, static (projectContexts, contextId) => projectContexts.Remove(contextId))));
-        }
+        // Pass-through for unit testing.
+        internal IDisposable RegisterContext(string contextId) => _activeEditorContextTracker.RegisterContext(contextId);
     }
 }
